Tinkering with bundling.
