--- index.html
@@ -1,40 +1,0 @@
-<html>
-    <head>
-        <title>Sample App</title>
-
-        <script src="build/jsweb/gnoll.js"></script>
-        <script>
-          var roll_full_options = null;
-
-          Module.onRuntimeInitialized = _ => {
-            roll_full_options = Module.cwrap('roll_full_options', 'number', ['string',
-              'string', 'number', 'number', 'number', 'number', 'number',
-              'number']);
-
-            //roll_full_options("1d100", null, 0, 0, 0, 1, 0, 0);
-          };
-
-          function roll(notation) {
-            Module.resetOut();
-            roll_full_options(notation, null, 0, 0, 0, 1, 0, 0);
-
-            return Module.stdout;
-          }
-
-          function onRoll() {
-            var notation = document.getElementById("tbNotation");
-            var output = document.getElementById("output");
-            var result = roll(notation.value);
-
-            output.innerHTML = result;
-          }
-        </script>
-    </head>
-    <body>
-      <div>
-        <p>Input: <input type="text" id="tbNotation" /></p>
-        <p><button onclick="onRoll()">Roll</button></p>
-      </div>
-      <div id="output"></div>
-    </body>
-</html>
