+++ src/js/package.json
